--- tests/misc/comments.c
@@ -1,18 +1,0 @@
-#include <stdio.h>
- 
-int main()
-{
-   // Single line comment in c source code
- 
-   printf("Writing comments is very useful.\n");
- 
-   /*
-    * Multi line comment syntax
-    * Comments help us to understand code later easily.
-    * Will you write comments while developing programs ?
-    */
- 
-   printf("Good luck c programmer.\n"); 
- 
-   return 0;
-}
