--- tests/misc/hello-world.c
@@ -1,7 +1,0 @@
-#include <stdio.h>
- 
-int main()
-{
-    printf("Hello World\n");
-    return 0;
-}
