--- tests/misc/if-else.c
@@ -1,13 +1,0 @@
-#include <stdio.h>
- 
-int main()
-{
-   int x = 2;
- 
-   if ( x == 1 )
-      printf("x is equal to one.\n");
-   else
-      printf("For comparison use == as = is the assignment operator.\n");
- 
-   return 0;
-}
