--- tests/misc/user-input.c
@@ -1,13 +1,0 @@
-#include <stdio.h>
- 
-int main()
-{
-   int number;
- 
-   printf("Enter an integer\n");
-   scanf("%d",&number);
- 
-   printf("Integer entered by you is %d\n", number);
- 
-   return 0;
-}
