--- tests/misc/while.c
@@ -1,14 +1,0 @@
-#include <stdio.h>
- 
-int main()
-{
-   int value = 1;
- 
-   while(value<=3)
-   {
-      printf("Value is %d\n", value);
-      value++;
-   }
- 
-   return 0;
-}
