+++ tests/user-input.c
@@ -1,0 +1,13 @@
+#include <stdio.h>
+
+int main()
+{
+    int number;
+
+    printf("Enter an integer\n");
+    scanf("%d", &number);
+
+    printf("Integer entered by you is %d\n", number);
+
+    return 0;
+}
