+++ tests/while.c
@@ -1,0 +1,14 @@
+#include <stdio.h>
+
+int main()
+{
+    int value = 1;
+
+    while (value <= 3)
+    {
+        printf("Value is %d\n", value);
+        value++;
+    }
+
+    return 0;
+}
